requirements.txt
fastapi
uvicorn
pandas
requests
psycopg2-binary